// src/modules/shared/enum/category.enum.ts
/* eslint-disable @typescript-eslint/no-duplicate-enum-values */
export enum CategoryEnum {
  // 624, 622, 647, 621, 644, 639, 642, 646
  // PERFUMARIA = 'PERFUMARIA',
  PERFUMES = 'PERFUMES', //622
  MAQUIAGENS = 'MAQUIAGENS', //647
  DERMOCOSMETICOS = 'DERMOCOSMETICOS', //621
  CABELO = 'LINHA CAPILAR', //644
  HIGIENE = 'HIGIENE&BELEZA', //639
  LEITE = 'LEITES', // 646
  FRALDAS = 'FRALDAS', // 646
  INFANTIL = 'INFANTIL' //646
}

export enum CategoryIdsEnum {
  // 624, 622, 647, 621, 644, 639, 642, 646
  // PERFUMARIA = '624',
  PERFUMES = '622',
  MAQUIAGENS = '647',
  DERMOCOSMETICOS = '621',
  CABELO = '644',
  HIGIENE = '639',
  LEITE = '642',
  FRALDAS = '646'
}

export enum CategorySlugEnum {
  'PERFUMARIA' = 'perfumaria',
  'PERFUMES' = 'perfumaria',
  'MAQUIAGENS' = 'maquiagens',
  'DERMOCOSMETICOS' = 'dermocosmeticos',
  'LINHA CAPILAR' = 'linha-capilar',
  'HIGIENE&BELEZA' = 'higienebeleza',
  'LEITES' = 'linha-infantil',
  'FRALDAS' = 'linha-infantil'
}